+++ .vscode/launch.json
@@ -5,7 +5,7 @@
       "name": "Launch",
       "type": "node",
       "request": "launch",
-      "program": "${workspaceRoot}/bin/tests/run.js",
+      "program": "${workspaceRoot}/bin/tests/mimc.js",
       "cwd": "${workspaceRoot}/bin",
       "preLaunchTask": "build",
       "env": {
