+++ bin/lib/ScriptSpecs.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const expressions_1 = require("./expressions");
+const StaticExpression_1 = require("./expressions/StaticExpression");
 const utils_1 = require("./utils");
 // CLASS DEFINITION
 // ================================================================================================
@@ -40,7 +40,7 @@ class ScriptSpecs {
             if (this.staticConstants.has(constant.name)) {
                 throw new Error(`Static constant '${constant.name}' is defined more than once`);
             }
-            let constExpression = new expressions_1.StaticExpression(constant.value, constant.name);
+            let constExpression = new StaticExpression_1.StaticExpression(constant.value, constant.name);
             this.staticConstants.set(constant.name, constExpression);
             if (utils_1.isMatrix(constant.dimensions)) {
                 this.constantBindings[constant.name] = this.field.newMatrixFrom(constant.value);
