+++ bin/lib/contexts/index.js
@@ -2,8 +2,8 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 var ExecutionContext_1 = require("./ExecutionContext");
 exports.ExecutionContext = ExecutionContext_1.ExecutionContext;
-var ProvingContext_1 = require("./ProvingContext");
-exports.ProofContext = ProvingContext_1.ProofContext;
+var ProofContext_1 = require("./ProofContext");
+exports.ProofContext = ProofContext_1.ProofContext;
 var VerificationContext_1 = require("./VerificationContext");
 exports.VerificationContext = VerificationContext_1.VerificationContext;
 //# sourceMappingURL=index.js.map
