--- bin/lib/expressions/index.js
@@ -1,7 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var Expression_1 = require("./Expression");
-exports.Expression = Expression_1.Expression;
-var StaticExpression_1 = require("./StaticExpression");
-exports.StaticExpression = StaticExpression_1.StaticExpression;
-//# sourceMappingURL=index.js.map
