+++ lib/ScriptSpecs.ts
@@ -4,7 +4,8 @@ import { StarkLimits } from '@guildofweavers/air-script';
 import { FiniteField } from '@guildofweavers/galois';
 import { ReadonlyRegisterSpecs, InputRegisterSpecs } from './registers';
 import { ReadonlyRegisterGroup, ConstantDeclaration } from './visitor';
-import { Expression, StaticExpression } from './expressions';
+import { Expression } from './expressions/Expression';
+import { StaticExpression } from './expressions/StaticExpression';
 import { isPowerOf2, isMatrix, isVector } from './utils';
 
 // CLASS DEFINITION
