--- lib/expressions/index.ts
@@ -1,2 +1,0 @@
-export { Expression } from './Expression';
-export { StaticExpression } from './StaticExpression';
